lang test

+++ src/ua/Endertainment/QuartzDefenders/Stats/StatsPlayer.java
@@ -22,7 +22,7 @@ public class StatsPlayer {
 	
 	private int wins;
 	private int games;
-	private float winRate;
+	private int winRate;
 	
 	private int level;
 	private int nextLevel;
@@ -47,7 +47,7 @@ public class StatsPlayer {
 		this.wins = StatsManager.getInfo(id, "wins"); //
 		this.games = StatsManager.getInfo(id, "games"); // 
 		if(games == 0) this.winRate = 0;
-		else this.winRate = ((float)wins / (float)games) * 100;
+		else this.winRate = (wins / games) * 100;
 		
 		this.brokenQuartz = StatsManager.getInfo(id, "quartz"); //
 		this.brokenOres = StatsManager.getInfo(id, "ores"); //
@@ -164,7 +164,7 @@ public class StatsPlayer {
 		this.games = games;
 		StatsManager.saveInfo(id, "games", this.games);
 	}
-	public float getWinRate() {
+	public int getWinRate() {
 		return winRate;
 	}
